صدا

+++ app/src/main/java/psb/com/kidpaint/webApi/paint/getLeaderShip/model/LeaderModel.java
@@ -33,7 +33,7 @@ public class LeaderModel implements Serializable
     private String url;
     @SerializedName("score")
     @Expose
-    private Double score;
+    private Integer score;
     private final static long serialVersionUID = -2499051059531575793L;
 
     public Integer getId() {
@@ -92,11 +92,11 @@ public class LeaderModel implements Serializable
         this.url = url;
     }
 
-    public Double getScore() {
+    public Integer getScore() {
         return score;
     }
 
-    public void setScore(Double score) {
+    public void setScore(Integer score) {
         this.score = score;
     }
